Actual numbers for first 15 upgrade types implemented \n Converted all int to long

+++ Assets/Scripts/StatTracker.cs
@@ -10,9 +10,9 @@ public class StatTracker : Singleton<StatTracker>
 
     private long timeStamp;
 
-    private int money;
-    private int moneyPerClick;
-    private int moneyRate;
+    private long money;
+    private long moneyPerClick;
+    private long moneyRate;
 
     void Awake()
     {
@@ -25,37 +25,37 @@ public class StatTracker : Singleton<StatTracker>
     // Private to prevent being called outside of singleton instance.
     private StatTracker() { }
 
-    public int getMoney()
+    public long getMoney()
     { 
         return money;
     }
 
-    public int getMoneyPerClick()
+    public long getMoneyPerClick()
     {
         return moneyPerClick;
     }
 
-    public int getMoneyRate()
+    public long getMoneyRate()
     {
         return moneyRate;
     }
 
 
-    public void incrementMoney(int x)
+    public void incrementMoney(long x)
     {
         money += x;
     }
 
-    public void decrementMoney(int x)
+    public void decrementMoney(long x)
     {
         money -= x;
     }
 
-    public void incrementMoneyPerClick(int x)
+    public void incrementMoneyPerClick(long x)
     {
         moneyPerClick += x;
     }
-    public void incrementeMoneyRate(int x)
+    public void incrementeMoneyRate(long x)
     {
         moneyRate += x;
     }
